Began implementing landing page
