+++ pages/index.tsx
@@ -9,15 +9,18 @@ import {
 	Heading,
 	HStack,
 	Image,
+	Img,
 	Input,
 	Link,
 	Spacer,
+	Text,
+	useColorMode,
 	useToast,
 	VStack,
 } from "@chakra-ui/react";
 import type { NextPage } from "next";
 import { useRouter } from "next/router";
-import React, { useState } from "react";
+import React, { useCallback, useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import {
 	auth,
@@ -29,170 +32,84 @@ import {
 import { ADD_USER } from "../gql/mutations/mutation";
 
 const Home: NextPage = () => {
+	const { colorMode, toggleColorMode } = useColorMode();
+
 	const router = useRouter();
 
-	const toast = useToast();
+	const onTryItClick = useCallback<
+		React.MouseEventHandler<HTMLButtonElement>
+	>(() => {
+		router.push("/login");
+	}, []);
 
-	const [email, setEmail] = useState("");
-	const [pass, setPass] = useState("");
-	const [invalid, setInvalid] = useState(false);
-	const [invalidText, setInvalidText] = useState("");
-
-	// We can ignore GQL errors, the only err
-	const [addUser] = useMutation(ADD_USER, { errorPolicy: "ignore" });
-
-	const [user, loading, err] = useAuthState(auth);
-
-	const resetState = () => {
-		setEmail("");
-		setPass("");
-		setInvalid(false);
-		setInvalidText("");
-	};
-
-	const onEmailChange = (e: React.FormEvent<HTMLInputElement>) => {
-		setEmail(e.currentTarget.value);
-	};
-
-	const onPassChange = (e: React.FormEvent<HTMLInputElement>) => {
-		setPass(e.currentTarget.value);
-	};
-
-	const onForgetPass = async () => {
-		if (email !== "")
-			try {
-				await sendPassReset(email);
-			} catch (e: any) {}
-		router.push("/passwordReset");
-	};
-
-	const onSubmit = async (e: React.FormEvent) => {
-		e.preventDefault();
-		try {
-			await logInWithEmailAndPass(email, pass);
-		} catch (e: any) {
-			setInvalid(true);
-			if (
-				e.code === "auth/wrong-password" ||
-				e.code === "auth/internal-error"
-			) {
-				setInvalidText("Incorrect password.");
-			}
-			return;
-		}
-
-		resetState();
-	};
-
-	const onSignInAnon = async (e: React.MouseEvent<HTMLButtonElement>) => {
-		e.preventDefault();
-
-		try {
-			var { uid, email: userEmail } = await logInAnonymously();
-		} catch (e: any) {
-			toast({
-				title: "Sign In Error",
-				description: `There was an issue signing in as guest.\nCode:  ${e}`,
-				status: "error",
-				duration: 5000,
-				isClosable: true,
-			});
-			return;
-		}
-
-		if (uid && userEmail) {
-			await addUser({ variables: { email: userEmail, userId: uid } });
-		}
-	};
-
-	const onSignInWithGoogle = async (e: React.MouseEvent<HTMLButtonElement>) => {
-		e.preventDefault();
-
-		try {
-			var { uid, email: userEmail } = await logInWithGoogle();
-		} catch (e: any) {
-			toast({
-				title: "Sign In Error",
-				description: `There was an issue signing in with google.\nCode:  ${e}`,
-				status: "error",
-				duration: 5000,
-				isClosable: true,
-			});
-			return;
-		}
-
-		if (uid && userEmail) {
-			await addUser({ variables: { email: userEmail, userId: uid } });
-		}
-	};
-
-	if (user) {
-		router.push("/dashboard");
-	}
+	const topBar = (
+		<HStack
+			bg={colorMode === "light" ? "white" : "gray.800"}
+			w="full"
+			padding={5}
+		>
+			<HStack>
+				<Box>
+					<Image
+						boxSize={["40px", "50px", "75px"]}
+						borderRadius="lg"
+						src="/android-chrome-192x192.png"
+						alt="apptrack-logo"
+					/>
+				</Box>
+				<Heading size="2xl">apptrack</Heading>
+			</HStack>
+			<Spacer />
+			<HStack>
+				<Button size="lg" w={["75px", "100px", "115px"]} variant="link">
+					About
+				</Button>
+				<Button
+					colorScheme="purple"
+					size="lg"
+					w={["75px", "100px", "115px"]}
+					onClick={onTryItClick}
+				>
+					Try It!
+				</Button>
+			</HStack>
+		</HStack>
+	);
 
 	return (
-		<Flex w="full" h="100vh" justify="center" align="center">
-			<Box boxShadow="lg" w="md" borderWidth="1px" borderRadius="lg" p={4}>
-				<form onSubmit={onSubmit}>
-					<VStack align="flex-start" w="full" h="full" spacing={4}>
-						<HStack>
-							<Image
-								borderRadius="lg"
-								boxSize="50px"
-								src="/android-chrome-192x192.png"
-							/>
-							<Heading>Apptrack</Heading>
-						</HStack>
-						<Heading>Log In</Heading>
-						<Spacer />
-						<Input value={email} placeholder="Email" onChange={onEmailChange} />
-						<FormControl isInvalid={invalid}>
-							<Input
-								isInvalid={invalid}
-								value={pass}
-								placeholder="Password"
-								type="password"
-								onChange={onPassChange}
-							/>
-							<FormErrorMessage>{invalidText}</FormErrorMessage>
-						</FormControl>
-						<Link color="purple.400" onClick={onForgetPass}>
-							Forgot password?
-						</Link>
-						<Spacer />
-						<Divider />
-						<Button colorScheme="purple" w="full" type="submit">
-							Sign in with email
-						</Button>
-						<Button
-							colorScheme="purple"
-							w="full"
-							variant="outline"
-							leftIcon={<Image src="/GOOGLE_G.png" boxSize="20px" />}
-							onClick={onSignInWithGoogle}
-						>
-							Sign in with Google
-						</Button>
-						<Button
-							colorScheme="purple"
-							w="full"
-							variant="outline"
-							onClick={onSignInAnon}
-						>
-							Try as Guest
-						</Button>
-						<Button
-							colorScheme="gray"
-							w="full"
-							onClick={() => {
-								router.push("/register");
-							}}
-						>
-							Create Account
-						</Button>
-					</VStack>
-				</form>
-			</Box>
+		<Flex
+			w="full"
+			h="100vh"
+			bg={colorMode === "light" ? "gray.50" : "gray.900"}
+			alignItems="start"
+			direction="column"
+		>
+			{topBar}
+			<Flex
+				direction="row"
+				justifyContent="space-around"
+				alignItems="center"
+				w="full"
+				h="full"
+			>
+				<Box bg="tomato" w="500px" h="500px"></Box>
+				<Divider h="600px" orientation="vertical" />
+				<Flex
+					w="500px"
+					h="500px"
+					display="flex"
+					flexDirection="column"
+					justifyContent="center"
+				>
+					<Text fontWeight="bold" fontSize="6xl" mb="20px">
+						Track it easy.
+					</Text>
+					<Text fontWeight="light" fontSize="4xl">
+						apptrack makes keeping tabs on all of your job applications super
+						simple.
+					</Text>
+				</Flex>
+			</Flex>
 		</Flex>
 	);
 };
